always install our desired version of controller-gen
The shell script to install controller-gen is a little verbose, so
move it out of the Makefile and into a separate hack script.

Have the script fetch the version of controller-tools that we expect,
then build it and install the results into the bin directory under the
root of the repository.

Have all of the commands that use controller-gen ensure that it is
built.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>