remove test dependency from docker build
There doesn't seem to be much point in running all of the lint and
unit tests before building the docker image, and it means we have to
keep the CI system set up to run extra tools in that job. This patch
removes the dependency so `make docker` runs the code generators and
then builds the image.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>