add a test for validating access to a node known to ironic
Show that when the host has no Provisioning.ID set we can find the node by name
and we link it to the host by setting the Provisioning.ID field.

Change the CreateNodes() method of the test server to take a callback
so the node "created" by the method can be passed back to the test for
checking.

Only allow POST calls to invoke the handler from CreateNodes().

Unmarshal the input to the CreateNodes() handler and set the UUID in
the node we get instead of changing the JSON string.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>