add test for validating access when the credentials have changed
Add a test to show that when we validate management access for a node
and think that the credentials have changed we try to update the
driver settings of the node.

Add NodeUpdate() method to support GET and PATCH of the same node URL.

Add a helper method to MockServer to make it easier to send a dynamic
JSON response.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>