Ensure registration in every state
Previously we only verified that the Node was registered in Ironic when
the credentials changed or in a steady state (Ready, Provisioned,
ExternallyProvisioned). Simplify the code by always calling
actionRegistering() prior to running the state machine.

This ensures that the Node is always created in the ironic database
after a move of the pod, regardless of what state the host is in at the
time.

This allows a BMC password to be changed in any order. Previously if the
credentials were updated in the Secret prior to changing them on the BMC
itself, it would never retry after the new credentials failed.

It also ensures that after an error caused by a temporary inability to
contact the BMC, retries still happen. Previously if there had been a
registration error, we just gave up because the credentials hadn't
changed on the incorrect assumption that bad credentials were the only
way of failing to register. (Although the Host would still be reconciled
every 1-2 minutes, it had no way of getting out of this state).

Fixes #708