Deprovision Node during deletion whenever possible
Previously if the Host had been marked for deletion and no corresponding
Node is registered in Ironic (as would happen e.g. after the ironic pod
moves to another k8s node), we wouldn't re-register it in order to
complete the deprovisioning. Instead we would simply continue and allow
the Host to be deleted. This was to ensure that if there were no longer
credentials available, we would not block deletion of the Host.

This change means that we will only skip deprovisioning and go straight
to deletion if there are, in fact, no credentials available.

If credentials are available but wrong, then the Host will be stuck
deprovisioning. A user finding themselves in this situation can either
correct the credentials (to permit deprovisioning) or remove them
altogether (e.g. by deleting the Secret) to skip to deletion.