Use provisioned image data for adoption
When registering a Host as a Node in ironic and setting the image
information, prefer the current image data from the Status to the
requested image from the Spec. Previously we looked only at the Spec.

The only time this is likely to make any difference is during
deprovisioning, when there may be no image specified in the Spec.