ironic: Don't do automated cleaning twice in a row
When we deprovision a node in Ironic, by setting the target state to
deleted (which has nothing to do with deleting the node itself), it goes
through automated cleaning and returns to the available state.

Previously we would then move the state from available to manageable
before the deprovisioning was considered complete. The consequence of
this was that to provision the host again, we would first have to move
it back to the available state, which involves a second automated
cleaning.

We now leave the node in the available state after deprovisioning, so
that the host will now go through automated cleaning only before the
first time it is provisioned and once between each provisioning. In
order to make this work, when provisioning we have to set the instance
info in the available state rather than in the manageable state.

The first time we provision, the host will still be in the manageable
state. To perform manual cleaning (which is required to change RAID
settings) also requires returning to the manageable state.