Don't adopt in deprovisioning if provisioning didn't complete
In the deprovisioning state, calling Adopt() on a freshly-reregistered
host will fail if we didn't previously complete provisioning because
there will be no image data stored in the status to pass to the node for
adoption.

Only call Adopt() if the previous provisioning completed and recorded
the requisite image data.

If the node has just been re-registered, it will be in the manageable
state anyway, and thus will still get cleaned before it is next
provisioned.