Delay retry on deprovision failed
If deprovisioning fails, we were just repeatedly retrying it instead of
setting an error. Instead, set an error and retry only after a backoff.

If we are deprovisioning because of a deletion, give up after 3 retries
(since this may indicate that the host simply doesn't exist any more)
and allow the Host to be deleted.