Handle dual-stack configuration in inspection data
Currently we assume only an ipv4 *or* an ipv6 IP exists in the
inspection data, but it's valid for both to exist in a dual-stack
scenario.

To avoid changing the BMH API, in this case we create an additional
nic entry in the list, which will contain each IP but otherwise
the same data.

Fixes: #458