Ironic: Add result functions
Add a set of functions that lend semantic meaning to the various Result
values that can be returned - some of which result in identical Result
structures being returned to the caller. This is the first step toward
improving the provisioner API to allow the controller more insight into
how it should respond to events in the provisioner.

The six possible events are:

* A conflict that should be retried with a delay
* A change to the Host Status that requires it to be written back to the
  k8s API
* Waiting for an ongoing process
* Successful completion of the current operation
* A failure of the current operation
* A transient error (e.g. network connection failure)