Fixture: store state separately to provisioner
The fixtureProvisioner object gets recreated for each Reconcile() call, so
it cannot be used to store state that is used to make different things
happen on each reconciliation. Some of this state was instead being stored
in the host Status, which could potentially confound some of the tests:
only the controller should modify the status, and it is the thing we are
trying to test.

Instead, create a Fixture structure that contains persistent state for a
particular Host and the provisioner Factory to create a provisioner that
has access to that state. Stop writing to the Host's Status from the
fixture provisioner.