Don't write Status on every actionContinue
When the provisioner could write to the Host's Status subresource and
expect it to be persisted, we had to always treat a Dirty flag returned
from the provisioner as indicating that the Status was dirty and required a
write. (Oddly this meant the Dirty flag was treated with almost the inverse
of its intended meaning: when it was returned true the Status was rarely
actually dirty, while when it was false the action was complete which
always requires a write.)

Now that provisioners are no longer allowed to write to the Status, we only
need to request a write when the controller has updated the Status. Add a
new actionUpdate to cover the couple of cases that are not already covered
by actionComplete, and those where ClearError() returns dirty. The
actionUpdate type comprises actionContinue so that type checks for the
latter handle it correctly.

Make actionContinue return false from the Dirty() method, meaning that it
will not cause a write to the k8s API. The actionContinueNoWrite type is
removed, since it would now be identical to actionContinue.