Ask Ironic to include NodeUUID when querying ports
When registering a new host, we check if its bootMACAddress conflicts
with any existing Ironic nodes.  This procedure queries Ironic ports,
and if a port is found, we use the port's NodeUUID to find the
conflicting node.  However, we are using `ports.List()` which only
returns minimal resources by default; e.g. any returned ports will have
their NodeUUID property set to an empty string.

We are now explicitly asking for Ironic to include the `node_uuid`
field.

We also adjust the mock test mechanism to account for this.