Add capabilities to BMO to check whether hard reboot is requested
The default reboot-interface behaviour is to attempt a soft power
off, and if this fails, revert to a hard power off (PR #294). For high
availability use-cases we require the ability to immediately power-off
a node. This PR attempts to address that requirement and is part of a
wider solution requiring the CAPBM to set the annotation that we have
detailed and implemented in this commit. The baseline provisioner API
changes have been provided in an earlier commit.

CAPBM PR: https://github.com/openshift/cluster-api-provider-baremetal/pull/138
Also see: https://bugzilla.redhat.com/show_bug.cgi?id=1927678