Pass BootMACAddress to provisioner in HostData
Although the ironic provisioner uses the BootMACAddress only during
registration, conceptually it can be regarded as part of the identity of
the Host and therefore potentially of use to the provisioner in any call
that it handles. Therefore pass it to the provisioner factory in the
HostData so it is always available to the provisioner.