Handle hardwaredetails annotation value error
Currently we only handle the case where the json fails to parse
e.g is invalid json, but we silently ignore any value which is
valid data but fails to unpack into the HardwareDetails object.

In this case we should avoid removing the annotation since it's
not been copied into the status.hardware, and return an error
so the user can detect that the json data wasn't in the expected
format.