Really delay retry on inspect failed
In the patch 2d02462fb45dbb5dd9dd0189ce91e0c417bbab2e we were supposed
to report an error and wait for an exponentially increasing backoff time
before retrying. However, due to a missing early return, we are retrying
immediately with the result that the error gets cleared in fairly short
order.

Return early on failure as intended. Also, use the operationFailed()
convenience function instead of ad hoc updates to the return variables;
this case was likely missed due to the lack of a return.