Always update image details
Always update the image data within ValidateManagementAccess()
regardless of Ironic node existence or not. Currently, image data
update is happening only if Ironic node doesn't exist. That means, if
we failed to write image data, we won't be able to update it again in
the next retries since node already exists.