Fix adoption in ExternallyProvisioned state
In 269ff1eef216af719251daa4a33bc294409d1945 we stopped potentially
making use of the image from the Spec in the provisioner when there is
no image in the Status, except for the Provisioning state where we want
to register with the image we are provisioning even though it hasn't
completed yet.

We also need to make an exception for the ExternallyProvisioned state,
because in that state the image details are never copied to the Status.

This relies on the change in c0373bbfa38d3169d1b994cf0f0429396f12e651,
which ensures that the image data can be added to the Node once the Host
reaches the ExternallyProvisioned state. Prior to this, the image data
was only added when the Node was created, so we would have had to pass
the image details starting with the first call to
ValidateManagementAccess() in the Registration state.