ironic: Update image options only when changed
Since c0373bbfa38d3169d1b994cf0f0429396f12e651 we set the image options
on every registration. This is necessary, because we may have failed to
set these options immediately after creating the node, and also because
the image options are no longer passed in the Registration state but may
be needed in the ExternallyProvisioned state. However, currently we do
not check for the presence of the correct option values, which means we
are doing an extra unconditional write to ironic on virutally every
reconcile.

Reduce the load on ironic by not setting the options unless we are
changing or adding values.