ironic: Pass nodeUpdater instead of updates list
The nodeUpdater contains a list of updates, so just pass a single one of
those around rather than creating a new one each time we have updates
and concatenating the result onto an external updates list.

Since we are changing function signatures, get rid of all the error
returns that are not used as well.