ironic: Do automated cleaning in Prepare state
After completing manual cleaning, move the Node to the available state
before returning the Host to the Ready state. Previously we left it in
manageable; this transition involves doing automated cleaning. This
should mean that a Host in the Ready/Available state should usually
(with a few exceptions) represent a Node in the available state.

On Hosts that are created with provisioning information specified, this
will have little real-world effect. However, on Hosts where provisioning
information is specified later, this should make provisioning quicker
because automated cleaning will already have been performed.