Refactor actionPreparing()
It's hard to follow what happens when we unconditionally update the
current provisioning settings at the start of the function, and then
restore the original current settings if we didn't start a new manual
cleaning.

Instead, get the new settings without writing them to the Host
structure, and only update the Host status in the case where we did
start a new manual cleaning.

Note that this relies on Prepare() always returning started=true at
least once. Previously, Prepare could (but no longer does) simply return
complete (Dirty=false) to imply that the settings should be updated.