Always restart Prepare after an error
We previously assumed that clearing the provisioning settings after a
PreparationError would always be sufficient to restart it. However, if
the settings in the Spec are also empty then this will not be
sufficient, and the state of the actual Host is unknown.

To guard against this, always restart the Prepare step when a
PreparationError is present, regardless of whether the settings in the
Spec and Status match. Don't clear the error until the provisioner
reports that the cleaning has actually started.