Return started=true from Prepare() at least once
Previously we assumed that if Prepare() was complete (i.e. Dirty=false)
that implied that it was time to save the host config to the status. We
were relying on this when there was no work to do in Prepare.

However, it is better in these circumstances to be explicit about when
the status can be saved. In future, determining that there are no manual
cleaning steps required will not be the last thing to happen in
Prepare(). We want to update the status as soon as possible.