Fix transitions to Preparing state from Ready/Available
We were unintentionally transitioning the Host from the Ready state to
Preparing whenever the power state changed. The assumption was that
actionUpdate was only returned when the host config had changed.
However, that is incorrect: it is also returned by manageHostPower().

Don't wait until provisioning is requested to return to the Preparing
state. We want to do it as soon as the host config is out of sync, to
make the provisioning process as short as possible.

Fixes #883