unit tests: Make test data more realistic
In hostBuilder, make the root device hints in the Status match the Spec
by default, so Preparing won't be triggered.

The Status Image URL is never set in the Ready state, so don't use that
as a way of forcing NeedsProvisioning() to return false. Instead, clear
the Image URL in the Spec.