Handle separate apis go module for controller-gen
Because apis is in a separate go module now, we need to run
controller-gen on this module not on the main one. We don't want to make
the apis module depend on controller-gen, because the whole point is to
minimise dependencies. However, we can't use "go run" to across modules
because we can't switch modules between build and run.

Accordingly, split the build and run phases so that we can keep the
version information in the main module but run in the apis module.

Fixes #875