Handle node "available" state in inspection
When an annotation appears to force re-inspection, the Host returns to
the Inspecting state from the Ready/Available state. Originally the
ironic node would always be in the "manageable" state for a ready Host,
but we have usually been leaving it in the "available" state after
deprovisioning since 18465130c8f42bda4aa8191b5093656dcc029d90, and after
the initial preparation since 9a55ce5bc63a2f24916f1f604c736570836fdaea.

Handle the "available" node state during inspection by moving the node
to the "manageable" state. Since the first call to InspectHardware may
not start a new inspection, we have to return a new value to indicate
when inspection has actually started and thus it is safe to clear the
annotation. This also fixes a bug where failing to set the boot mode due
to a conflict would cause the inspection annotation to be removed
without starting a new inspection.