Increase the default $BMO_CONCURRENCY for scale
Instead of defaulting the $BMO_CONCURRENCY value (the maximum number of
Hosts to reconcile concurrently) to a hard-coded value of 3, set it
instead to the number of CPU threads available, but constrained to a
range between 2 and 8.

We never default to 1 so that we don't inadvertantly rely on
single-threadedness in tests. 8 seems to be a reasonable value for a
large scale deployment, while still being substantially below the
default $PROVISIONING_LIMIT of 20.

There remains no restriction on the value that can be passed in the
environment variable.

Fixes #905

(cherry picked from commit febccb36e26181edc38c5fa894a5dcfec7eb4cb3)
Signed-off-by: Honza Pokorny <honza@redhat.com>