Don't clear BMC settings from Status after deprovisioning
The BootMode and the Hardware RAID config are set in the BMC, not by
software running on the Host, so they are not affected by
deprovisioning. Whatever values we configured during the previous
Prepare can be considered to still be in effect; there is no need to
initiate a new Prepare if the Spec has not changed.

The RootDeviceHints are not used anywhere until provisioning time
anyway, so they don't need to be cleared either.

The only exception is the Software RAID config, which cannot be relied
upon to still be in place and valid, and which Ironic clears in the
process of deprovisioning. We continue to reset this so that if software
RAID is still configured, we will initiate another Prepare to set it up.

Discussed with the Ironic team here:
https://meetings.opendev.org/irclogs/%23openstack-ironic/%23openstack-ironic.2021-06-09.log.html#t2021-06-09T17:42:31