Don't pass RootDeviceHints to provisioner Prepare()
The root device hints are not something we have to set up in the BMC,
just some data we need copied to the status before the host starts
provisioning. previously we passed them to the provisioner's Prepare
method, but this had the side-effect of forcing a new manual cleaning
and reset of the RAID settings every time the root device hints change.

Instead, just copy the RootDeviceHints directly whenever they differ,
and only once they are in sync check for changes that would trigger a
call to the provisioner's Prepare().