Set target RAID data in ironic node after DB lost
Ensure that if the ironic database is dropped after RAID is configured,
we re-add the RAID configuration to the node in ironic. This doesn't
matter for hardware RAID, but for software RAID it might be necessary
for ironic to manage it correctly.

This only sets the *Target* RAID config, but after cleaning we might
need to set the *current* RAID config as well, e.g. so that ironic knows
there is Software RAID set up.