Make check for detached annotation more explicit
Currently checking if the host is detached checks whether the
"baremetalhost.metal3.io/detached" annotation exists, but ignores its
value. This may cause some confusion in case we end up with a BMH
resource with the annotation value set to "false" - from the linguistic
perspective such a host is not detached, however for the current code
such a host would be treated as detached.

This PR makes it explicit so that only when the value of the annotation
is "true", we treat the host as detached.