ironic: Clear HW RAID configuration only when required
The Hardware RAID configuration only needs to be cleared when:

- There is a non-nil (but possibly empty) list of HW RAID volumes that
  differs from the existing configured RAID volumes; or
- There is no HW RAID configured, but there were existing configured HW
  RAID volumes.

Note that after an error in Preparing, we reset the RAID config in the
status back to nil, since we don't know the state the hardware is in. If
the user responds by setting the RAID settings back to nil then we leave
the hardware in whatever state we found it, to ensure that if the
failure was due to HW RAID not being supported, the user has a way to
recover. If HW RAID is supported and the user wants to clear it, they
need to explicitly set an empty list. To improve this we'd need to
remember more of the history: whether configuring HW RAID on the host
has _ever_ succeeded.