Use recreate strategy for ironic deployment
We cannot run Ironic with more than 1 replica (the second crashes
immediately) but the default rolling update strategy will try to get a
second replica up and running before it removes the old one. This means
that the rollout always gets stuck unless we first manually scale down
the deployment to 0, then update, and finally scale it up again.
The recreate strategy is a better fit since it does the scale down
automatically.