Create SecretManager type to manage Secrets
We currently have two implementations of adding Secrets to the cache,
and more will soon be required.

Create a SecretManager struct in a new secretutils package that can
handle fetching a Secret, adding a Label to ensure it is cached, and
optionally adding an Owner (which may be a Controller) to ensure it can
trigger reconciliation. This single implementation should be usable for
everything.