Fix missing case of BuildRAIDCleanSteps
When `hardwareRAIDVolumes` is nil and the target node doesn't have a RAID
controller. We shouldn't delete the RAID configuration on that node in
that case, otherwise the `ironic` node will enter the clean failed state
because of `delete_configuration` clean step failure. So in this case, hardware
RAID operations will be skipped.

Signed-off-by: Zou Yu <zouy.fnst@cn.fujitsu.com>