Add LIVE_ISO_FORCE_PERSISTENT_BOOT_DEVICE variable
When using format: live-iso, two distinct use-cases exist:
  - Boot an installer ISO (non-persistent one-time boot desired)
  - Boot an ISO appliance (persistent live-iso boot desired).

The original plan was to solve the installer-iso boot order via
uefibootmgr (running from the live-iso), but this doesn't work
reliably on all hardware.

So expose a variable to control this behavior - in future we may
want to consider a per-BMH variable, but for now expose a
per-BMO variable.