ironic: Refactor changePower() to publish events directly
If we publish the event on success inside changePower, we don't have to
have three different methods calling it that all have to check the
return values for success.

This allows us to eliminate the HostLockedError type.

It also fixes a bug where on Power Off we were publishing an event
saying the power was turned off in the case where we actually were
waiting for a change in the provisioning state.