gitignore: Get rid of rules not associated with the project
Everybody has their own set of tools that generate different tempfiles
that should be ignored. The correct place to configure this is in the
global gitignore file "~/.config/git/ignore". Including ignore rules for
a select set of tools is an antipattern that encourages people to want
to add their own tools to the list in each of the hundreds of projects
they interact with instead of setting up their global rules once.