Remove provisioning IP from ironic-dnsmasq checks
The liveness and readiness healthchecks for the ironic-dnsmasq container
were using the $PROVISIONING_IP to determine whether the dnsmasq process
was up and running properly. However, experimentally we've verified that
the dnsmasq service can serve on port 69 bound to a different IP on the
same subnet.. It should be sufficient to check that some service is
listening on port 69 for a sanity readiness / liveness check.