Fix software RAID
The current code treats software RAID similarly to hardware RAID and
expects the driver's RAIDInterface value to be "agent" to use it.
This is never the case, as software RAID is not driver-specific.

This change adjusts the procedure to set the expected RAIDInterface
before setting the target configuration. This makes software RAID work
with any values of the driver's RAIDInterface.

Note that there is still a potential issue with root device hints,
maybe the users should skip them when using software RAID.