Add make targets for setting images
These make targets can be used to change the image used for BMO, Ironic,
Mariadb, keepalived and IPA downloader. This is similar to how things work in CAPM3.

The deploy target was removed. It has been broken for at least a year
since the name of the image was changed. I.e. it could apply the
manifests but the image substitution didn't work.