Make bmc subscription validator less aggressive
Our intent is to make the subscription CR immutable; however, it doesn't
work if finalizers can't be added, or the status updated.  Therefore,
this patch loosens the immutability requirement to the spec only.  This
means that any k8s bookkeeping can still happen, and we can prevent
users from changing existing subscriptions.