Handle maintenance more gracefully
First, do not set the maintenance flag for nodes in manageable state,
it's not required.

Second, if the maintenance flag is on when we don't expect it and Ironic
does not report a fault, move the nodes off maintenance.

(cherry picked from commit 0e45cabd877445f6a7b2f2f1004d2072efb97c02)