Set minReadySeconds for BMO and Ironic
This is to make it easier to tell when the deployments are "available".
The default is to consider them available as soon as the pods have
started. But if there is some problem it is likely that the pods will
immediately crash. Therefore it makes sense to give them a few seconds
before we consider them available for real.