Allow image provider to override kernel and add kernel params
Currently, we assume that the same kernel can be used for all images.
This is not the case in multi-arch scenarios. Add a new optional field
to pass a kernel from the provider.

Also add support for extra kernel parameters. These are required by
CoreOS to link to the correct rootfs.

Requires: https://review.opendev.org/c/openstack/ironic/+/837379