Get rid of envsubst for nameprefix
The Ironic kustomizations used envsubst to set the nameprefix, while BMO
was using the built in feature in kustomize. To unify them and simplify
the deploy script a bit, the Ironix kustomizations were changed to also
use the kustomize feature.

There is a slight change in the manifests since the namePrefix is
applied also to secrets, configmaps and label selectors that didn't have
the prefix set before.

The namespace is now also created as part of the kustomizations for
Ironic (same as for BMO).