Don't register in Inspecting if inspection disabled
When inspection is disabled, the host still passes through the
inspecting state. Previously, we would still try to register the host in
this state, which would mean that it would require a valid
PreprovisioningImage to continue. Since no image is needed if we are not
actually going to inspect, skip registration in this case.