Don't require image build for RHACM <= 2.4
If registerHost() is called in the Preparing state (or the Inspecting
state), we will require a PreprovisioningImage to be built before
anything can proceed. The machine-image-customization-controller will
not reconcile PreprovisioningImage resources where the host has an
InfraEnv label from the Infrastructure operator in RHACM/MCE.

In RHACM <= 2.4 there was no controller provided for
PreprovisioningImage resources with the label. To prevent being unable
to use hosts with RHACM <= 2.4, do not call registerHost() in the
Preparing state for these hosts if inspection is disabled. We know that
these hosts never need to use features in the Preparing state, so this
is safe.

Later versions of RHACM/MCE will include a controller to reconcile the
image resources. They may also use features of Preparing and require a
pre-provisioning image to be available. However, if this is the case
then inspection will never be disabled.

See https://bugzilla.redhat.com/show_bug.cgi?id=2087213