add explicit securitycontexts to controller
It is always good to not rely on the defaults, but be explicit.
Set explicit, secure securityContext for the BMO controller manager
deployment and containers.

CAPI has the same starting from upcoming v1.4.0 and cert-manager etc
has them already.

Setting explicit securityContext has its downsides as well, for tilt.
Tilt's live update cannot handle securityContext which sets the user
as non-root, as it requires root to deploy the binaries on rebuild.
To workaround this, strip function is added to Tiltfile to handle
BMO securityContexts.